requirements.txt
flask
python-dotenv
google-cloud-storage
google-cloud-aiplatform
pandas